LayoutTests/webaudio/AudioParam/exponential-ramp-crash.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Test if a corner case crashes the exponential ramp.
    </title>
    <script src="../../imported/w3c/web-platform-tests/resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
    <script src="../resources/audit.js"></script>
  </head>
  <body>
    <script id="layout-test-code">
      const t = async_test('exponential-ramp-crash');
      const onload = () => {
        const context = new OfflineAudioContext(2, 441000, 44100);
        const source = new ConstantSourceNode(context);
        const delay_node = context.createDelay(30);
        delay_node.connect(context.destination);
  
        delay_node.delayTime.exponentialRampToValueAtTime(2, 4.1);
        delay_node.delayTime.cancelAndHoldAtTime(4);
        context.oncomplete = t.step_func_done(() => {
          assert_equals(delay_node.delayTime.value, 0);
          assert_equals(context.state, 'closed');
        });
        context.startRendering();
      };
      window.addEventListener('load', t.step_func(onload));
    </script>
  </body>
</html>
